validateLocation geliştirildi
