+++ DT_OutputActions/DT_OutputActions.iml
@@ -16,7 +16,7 @@
     </content>
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="jdk" jdkName="1.7" jdkType="JavaSDK" />
-    <orderEntry type="module-library" exported="">
+    <orderEntry type="module-library">
       <library name="sdk.jar">
         <CLASSES>
           <root url="jar://$MODULE_DIR$/lib/sdk.jar!/" />
@@ -25,6 +25,6 @@
         <SOURCES />
       </library>
     </orderEntry>
-    <orderEntry type="library" exported="" name="slf4j-api-1.7.21" level="project" />
+    <orderEntry type="library" name="slf4j-api-1.7.21" level="project" />
   </component>
 </module>
