+++ DT_OutputActions/locations.txt
@@ -1,0 +1,244 @@
+42.546245	1.601554
+23.424076	53.847818
+33.93911	67.709953
+17.060816	-61.796428
+18.220554	-63.068615
+41.153332	20.168331
+40.069099	45.038189
+12.226079	-69.060087
+-11.202692	17.873887
+-75.250973	-0.071389
+-38.416097	-63.616672
+-14.270972	-170.132217
+47.516231	14.550072
+-25.274398	133.775136
+12.52111	-69.968338
+40.143105	47.576927
+43.915886	17.679076
+13.193887	-59.543198
+23.684994	90.356331
+50.503887	4.469936
+12.238333	-1.561593
+42.733883	25.48583
+25.930414	50.637772
+-3.373056	29.918886
+9.30769		2.315834
+32.321384	-64.75737
+4.535277	114.727669
+-16.290154	-63.588653
+-14.235004	-51.92528
+25.03428	-77.39628
+27.514162	90.433601
+-54.423199	3.413194
+-22.328474	24.684866
+53.709807	27.953389
+17.189877	-88.49765
+56.130366	-106.346771
+-12.164165	96.870956
+-4.038333	21.758664
+6.611111	20.939444
+-0.228021	15.827659
+46.818188	8.227512
+7.539989	-5.54708
+-21.236736	-159.777671
+-35.675147	-71.542969
+7.369722	12.354722
+35.86166	104.195397
+4.570868	-74.297333
+9.748917	-83.753428
+21.521757	-77.781167
+16.002082	-24.013197
+-10.447525	105.690449
+35.126413	33.429859
+49.817492	15.472962
+51.165691	10.451526
+11.825138	42.590275
+56.26392	9.501785
+15.414999	-61.370976
+18.735693	-70.162651
+28.033886	1.659626
+-1.831239	-78.183406
+58.595272	25.013607
+26.820553	30.802498
+24.215527	-12.885834
+15.179384	39.782334
+40.463667	-3.74922
+9.145		40.489673
+61.92411	25.748151
+-16.578193	179.414413
+-51.796253	-59.523613
+7.425554	150.550812
+61.892635	-6.911806
+46.227638	2.213749
+-0.803689	11.609444
+55.378051	-3.435973
+12.262776	-61.604171
+42.315407	43.356892
+3.933889	-53.125782
+49.465691	-2.585278
+7.946527	-1.023194
+36.137741	-5.345374
+71.706936	-42.604303
+13.443182	-15.310139
+9.945587	-9.696645
+16.995971	-62.067641
+1.650801	10.267895
+39.074208	21.824312
+-54.429579	-36.587909
+15.783471	-90.230759
+13.444304	144.793731
+11.803749	-15.180413
+4.860416	-58.93018
+31.354676	34.308825
+22.396428	114.109497
+-53.08181	73.504158
+15.199999	-86.241905
+45.1		15.2
+18.971187	-72.285215
+47.162494	19.503304
+-0.789275	113.921327
+53.41291	-8.24389
+31.046051	34.851612
+54.236107	-4.548056
+20.593684	78.96288
+-6.343194	71.876519
+33.223191	43.679291
+32.427908	53.688046
+64.963051	-19.020835
+41.87194	12.56738
+49.214439	-2.13125
+18.109581	-77.297508
+30.585164	36.238414
+36.204824	138.252924
+-0.023559	37.906193
+41.20438	74.766098
+12.565679	104.990963
+-3.370417	-168.734039
+-11.875001	43.872219
+17.357822	-62.782998
+40.339852	127.510093
+35.907757	127.766922
+29.31166	47.481766
+19.513469	-80.566956
+48.019573	66.923684
+19.85627	102.495496
+33.854721	35.862285
+13.909444	-60.978893
+47.166		9.555373
+7.873054	80.771797
+6.428055	-9.429499
+-29.609988	28.233608
+55.169438	23.881275
+49.815273	6.129583
+56.879635	24.603189
+26.3351		17.228331
+31.791702	-7.09262
+43.750298	7.412841
+47.411631	28.369885
+42.708678	19.37439
+-18.766947	46.869107
+7.131474	171.184478
+41.608635	21.745275
+17.570692	-3.996166
+21.913965	95.956223
+46.862496	103.846656
+22.198745	113.543873
+17.33083	145.38469
+14.641528	-61.024174
+21.00789	-10.940835
+16.742498	-62.187366
+35.937496	14.375416
+-20.348404	57.552152
+3.202778	73.22068
+-13.254308	34.301525
+23.634501	-102.552784
+4.210484	101.975766
+-18.665695	35.529562
+-22.95764	18.49041
+-20.904305	165.618042
+17.607789	8.081666
+-29.040835	167.954712
+9.081999	8.675277
+12.865416	-85.207229
+52.132633	5.291266
+60.472024	8.468946
+28.394857	84.124008
+-0.522778	166.931503
+-19.054445	-169.867233
+-40.900557	174.885971
+21.512583	55.923255
+8.537981	-80.782127
+-9.189967	-75.015152
+-17.679742	-149.406843
+-6.314993	143.95555
+12.879721	121.774017
+30.375321	69.345116
+51.919438	19.145136
+46.941936	-56.27111
+-24.703615	-127.439308
+18.220833	-66.590149
+31.952162	35.233154
+39.399872	-8.224454
+7.51498		134.58252
+-23.442503	-58.443832
+25.354826	51.183884
+-21.115141	55.536384
+45.943161	24.96676
+44.016521	21.005859
+61.52401	105.318756
+-1.940278	29.873888
+23.885942	45.079162
+-9.64571	160.156194
+-4.679574	55.491977
+12.862807	30.217636
+60.128161	18.643501
+1.352083	103.819836
+-24.143474	-10.030696
+46.151241	14.995463
+77.553604	23.670272
+48.669026	19.699024
+8.460555	-11.779889
+43.94236	12.457777
+14.497401	-14.452362
+5.152149	46.199616
+3.919305	-56.027783
+0.18636		6.613081
+13.794185	-88.89653
+34.802075	38.996815
+-26.522503	31.465866
+21.694025	-71.797928
+15.454166	18.732207
+-49.280366	69.348557
+8.619543	0.824782
+15.870032	100.992541
+38.861034	71.276093
+-8.967363	-171.855881
+-8.874217	125.727539
+38.969719	59.556278
+33.886917	9.537499
+-21.178986	-175.198242
+38.963745	35.243322
+10.691803	-61.222503
+-7.109535	177.64933
+23.69781	120.960515
+-6.369028	34.888822
+48.379433	31.16558
+1.373333	32.290275
+37.09024	-95.712891
+-32.522779	-55.765835
+41.377491	64.585262
+41.902916	12.453389
+12.984305	-61.287228
+6.42375		-66.58973
+18.420695	-64.639968
+18.335765	-64.896335
+14.058324	108.277199
+-15.376706	166.959158
+-13.768752	-177.156097
+-13.759029	-172.104629
+42.602636	20.902977
+15.552727	48.516388
+-12.8275	45.166244
+-30.559482	22.937506
+-13.133897	27.849332
+-19.015438	29.154857
